chore: move helper function to another file
