+++ core/src/plugins/kubernetes/container/run.ts
@@ -9,11 +9,11 @@
 import type { ContainerRunAction } from "../../container/moduleConfig.js"
 import { runAndCopy } from "../run.js"
 import type { KubernetesPluginContext } from "../config.js"
-import { makePodName } from "../util.js"
+import { makePodName, toActionStatus } from "../util.js"
 import { getNamespaceStatus } from "../namespace.js"
 import type { RunActionHandler } from "../../../plugin/action-types.js"
 import { getDeployedImageId } from "./util.js"
-import { composeCacheableResult, toActionStatus } from "../results-cache-base.js"
+import { composeCacheableResult } from "../results-cache-base.js"
 import { getResultCache } from "../results-cache.js"
 
 export const k8sContainerRun: RunActionHandler<"run", ContainerRunAction> = async (params) => {
