+++ core/src/plugins/kubernetes/container/test.ts
@@ -8,12 +8,12 @@
 
 import type { ContainerTestAction } from "../../container/moduleConfig.js"
 import { runAndCopy } from "../run.js"
-import { makePodName } from "../util.js"
+import { makePodName, toActionStatus } from "../util.js"
 import { getNamespaceStatus } from "../namespace.js"
 import type { KubernetesPluginContext } from "../config.js"
 import type { TestActionHandler } from "../../../plugin/action-types.js"
 import { getDeployedImageId } from "./util.js"
-import { composeCacheableResult, toActionStatus } from "../results-cache-base.js"
+import { composeCacheableResult } from "../results-cache-base.js"
 import { getResultCache } from "../results-cache.js"
 
 export const k8sContainerTest: TestActionHandler<"run", ContainerTestAction> = async (params) => {
