+++ core/src/plugins/kubernetes/helm/helm-pod.ts
@@ -15,13 +15,13 @@ import type { CommonRunParams } from "../../../plugin/handlers/Run/run.js"
 import type { KubernetesPluginContext } from "../config.js"
 import { getActionNamespaceStatus } from "../namespace.js"
 import { k8sGetRunResult } from "../run-results.js"
-import { getResourceContainer, getResourcePodSpec, getTargetResource, makePodName } from "../util.js"
+import { getResourceContainer, getResourcePodSpec, getTargetResource, makePodName, toActionStatus } from "../util.js"
 import type { HelmPodRunAction, HelmPodTestAction } from "./config.js"
 import { helmPodRunSchema } from "./config.js"
 import { runAndCopy } from "../run.js"
 import { filterManifests, prepareManifests, prepareTemplates } from "./common.js"
 import { kubernetesRunOutputsSchema } from "../kubernetes-type/config.js"
-import { composeCacheableResult, toActionStatus } from "../results-cache-base.js"
+import { composeCacheableResult } from "../results-cache-base.js"
 import { getResultCache } from "../results-cache.js"
 
 const helmRunPodOutputsSchema = kubernetesRunOutputsSchema
