+++ core/src/plugins/kubernetes/kubernetes-type/kubernetes-pod.ts
@@ -31,8 +31,9 @@ import { kustomizeSpecSchema } from "./kustomize.js"
 import type { ObjectSchema } from "@hapi/joi"
 import type { TestAction, TestActionConfig } from "../../../actions/test.js"
 import { k8sGetTestResult } from "../test-results.js"
-import { composeCacheableResult, toActionStatus } from "../results-cache-base.js"
+import { composeCacheableResult } from "../results-cache-base.js"
 import { getResultCache } from "../results-cache.js"
+import { toActionStatus } from "../util.js"
 
 // RUN //
