+++ core/src/plugins/kubernetes/run-results.ts
@@ -7,8 +7,8 @@
  */
 
 import type { RunActionHandler } from "../../plugin/action-types.js"
-import { toActionStatus } from "./results-cache-base.js"
 import { getResultCache } from "./results-cache.js"
+import { toActionStatus } from "./util.js"
 
 // TODO: figure out how to get rid of the any cast here
 export const k8sGetRunResult: RunActionHandler<"getResult", any> = async (params) => {
